pyproject.toml
[tool.poetry]
name = "squeeze-alexa"
version = "2.5.0"
description = "Squeezebox integration for Amazon Alexa "
authors = ["Nick Boultbee <nick@declension.net>"]

[tool.poetry.dependencies]
python = "^3.6"
fuzzywuzzy = "=0.17"
paho-mqtt = "^1.4"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
coveralls = "^1.5"
flake8 = "^3.6"
pytest-cov = "^2.6"
pytest-sugar = "^0.9.2"
hbmqtt = "^0.9.5"
boto3 = "^1.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
